% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob-cal_slope.r
\name{cal_slope}
\alias{cal_slope}
\alias{cal_slope.data.frame}
\alias{cal_slope_vec}
\title{Calibration slope}
\usage{
cal_slope(data, ...)

\method{cal_slope}{data.frame}(
  data,
  truth,
  ...,
  estimator = NULL,
  na_rm = TRUE,
  case_weights = NULL,
  event_level = "first"
)

cal_slope_vec(
  truth,
  estimate,
  estimator = NULL,
  na_rm = TRUE,
  event_level = "first",
  case_weights = NULL
)
}
\description{
\code{cal_slope()} is a metric that computes the calibration slope.
}
\details{
The calibration slope evaluates the spread of the estimated risks and has a
target value of 1. A slope < 1 suggests that estimated risks are too extreme,
i.e., too high for patients who are at high risk and too low for patients
who are at low risk. A slope > 1 suggests the opposite, i.e., that risk
estimates are too moderate.
}
\references{
Van Calster, B., McLernon, D.J., van Smeden, M. et al. Calibration:
the Achilles heel of predictive analytics.
BMC Med 17, 230 (2019). https://doi.org/10.1186/s12916-019-1466-7
}
\seealso{
Other class probability metrics: 
\code{\link{cal_intercept}()}
}
\author{
An Tang
}
\concept{class probability metrics}
